requirements.txt
streamlit==1.35.0
PyPDF2==3.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
transformers==4.41.2
accelerate==0.30.1
bitsandbytes==0.43.1
huggingface-hub==0.23.3
langchain==0.1.17
langchain-core==0.1.48  # Must be >=0.1.48 for langchain 0.1.17
langchain-text-splitters==0.0.2  # Required for text splitting
torch==2.3.0